+++ parallax/src/camera/Camera2D.cpp
@@ -12,8 +12,8 @@ namespace prx {
 
 	Camera2D::~Camera2D() {}
 	
-	void Camera2D::init(float viewSpaceWidth, float viewSpaceHeight) {
-		m_ViewSpaceSize = hpm::vec2(viewSpaceWidth, viewSpaceHeight);
+	void Camera2D::init(unsigned viewSpaceWidth, unsigned viewSpaceHeight) {
+		m_ViewSpaceSize = hpm::vec2(static_cast<float>(viewSpaceWidth), static_cast<float>(viewSpaceHeight));
 		m_ProjectionMatrix = hpm::mat4::ortho(m_Position.x, m_ViewSpaceSize.x + m_Position.x,
 											  m_ViewSpaceSize.y + m_Position.y, m_Position.y,
 											  static_cast<float>(Director::maxDepthValue()),
